+++ core/src/test/java/org/gbif/vocabulary/persistence/mappers/VocabularyMapperTest.java
@@ -67,20 +67,20 @@ public class VocabularyMapperTest extends BaseMapperTest<Vocabulary> {
     vocabularyGbif.setNamespace("namespace gbif");
     vocabularyMapper.create(vocabularyGbif);
 
-    assertList("vocab1", null, null, null, Integer.MAX_VALUE,0);
-    assertList("vocab1", null, null, null, vocabulary1.getKey(),1);
-    assertList(null, null, null, null, vocabulary2.getKey(),1);
-    assertList("voc", null, null, null, null,3);
-    assertList("ocab", null, null, null, null,0);
-    assertList("namesp gb", null, null, null, null,1);
-    assertList(null, "vocab1", null, null, null,1);
-    assertList(null, "vocab", null, null, null,0);
-    assertList(null, null, "namespace gbif", null, null,1);
-    assertList(null, null, "namespace", null, null,0);
-    assertList(null, "vocab2", "namespace2", null, null,1);
-    assertList("voc", "vocab1", "namespace1", null, null,1);
-    assertList("oca", "vocab2", "namespace2", null, null,0);
-    assertList("v gbif", "vocab gbif", null, null, null,1);
+    assertList("vocab1", null, null, null, Long.MAX_VALUE, 0);
+    assertList("vocab1", null, null, null, vocabulary1.getKey(), 1);
+    assertList(null, null, null, null, vocabulary2.getKey(), 1);
+    assertList("voc", null, null, null, null, 3);
+    assertList("ocab", null, null, null, null, 0);
+    assertList("namesp gb", null, null, null, null, 1);
+    assertList(null, "vocab1", null, null, null, 1);
+    assertList(null, "vocab", null, null, null, 0);
+    assertList(null, null, "namespace gbif", null, null, 1);
+    assertList(null, null, "namespace", null, null, 0);
+    assertList(null, "vocab2", "namespace2", null, null, 1);
+    assertList("voc", "vocab1", "namespace1", null, null, 1);
+    assertList("oca", "vocab2", "namespace2", null, null, 0);
+    assertList("v gbif", "vocab gbif", null, null, null, 1);
   }
 
   @Test
@@ -94,13 +94,13 @@ public class VocabularyMapperTest extends BaseMapperTest<Vocabulary> {
     vocabularyMapper.create(vocabulary2);
 
     assertList(null, null, null, true, null, 0);
-    assertList(null, null, "n1", true, null,0);
-    assertList(null, null, "n1", null, null,2);
-    assertList(null, null, "n1", false, null,2);
+    assertList(null, null, "n1", true, null, 0);
+    assertList(null, null, "n1", null, null, 2);
+    assertList(null, null, "n1", false, null, 2);
 
     vocabularyMapper.deprecate(vocabulary1.getKey(), DEPRECATED_BY, null);
-    assertList(null, null, null, true, null,1);
-    assertList(null, null, "n1", true, null,1);
+    assertList(null, null, null, true, null, 1);
+    assertList(null, null, "n1", true, null, 1);
   }
 
   @Test
@@ -172,7 +172,12 @@ public class VocabularyMapperTest extends BaseMapperTest<Vocabulary> {
   }
 
   private void assertList(
-      String query, String name, String namespace, Boolean deprecated, Integer key, int expectedResult) {
+      String query,
+      String name,
+      String namespace,
+      Boolean deprecated,
+      Long key,
+      int expectedResult) {
     assertEquals(
         expectedResult,
         vocabularyMapper.list(query, name, namespace, deprecated, key, DEFAULT_PAGE).size());
